refactor: move isyntaxfactory interface into spi
