+++ core/src/main/java/dk/gtz/graphedit/internal/DemoSyntaxFactory.java
@@ -7,9 +7,9 @@ import java.util.UUID;
 import dk.gtz.graphedit.model.ModelEdge;
 import dk.gtz.graphedit.model.ModelVertex;
 import dk.gtz.graphedit.model.migration.ISyntaxMigrater;
+import dk.gtz.graphedit.spi.ISyntaxFactory;
 import dk.gtz.graphedit.tool.IToolbox;
 import dk.gtz.graphedit.view.EdgeController;
-import dk.gtz.graphedit.view.ISyntaxFactory;
 import dk.gtz.graphedit.view.ModelEditorController;
 import dk.gtz.graphedit.view.VertexController;
 import dk.gtz.graphedit.viewmodel.ViewModelEdge;
