+++ core/src/main/java/dk/gtz/graphedit/view/EdgeController.java
@@ -7,6 +7,7 @@ import org.slf4j.LoggerFactory;
 
 import dk.gtz.graphedit.tool.ITool;
 import dk.gtz.graphedit.events.EdgeMouseEvent;
+import dk.gtz.graphedit.spi.ISyntaxFactory;
 import dk.gtz.graphedit.util.BindingsUtil;
 import dk.gtz.graphedit.viewmodel.ViewModelEdge;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
