+++ core/src/main/java/dk/gtz/graphedit/view/VertexController.java
@@ -7,6 +7,7 @@ import org.slf4j.LoggerFactory;
 
 import dk.gtz.graphedit.tool.ITool;
 import dk.gtz.graphedit.events.VertexMouseEvent;
+import dk.gtz.graphedit.spi.ISyntaxFactory;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
 import dk.gtz.graphedit.viewmodel.ViewModelGraph;
 import dk.gtz.graphedit.viewmodel.ViewModelVertex;
