+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/lts/LTSSyntaxFactory.java
@@ -11,8 +11,8 @@ import dk.gtz.graphedit.plugins.syntaxes.lts.view.StateController;
 import dk.gtz.graphedit.plugins.syntaxes.lts.view.TransitionController;
 import dk.gtz.graphedit.plugins.syntaxes.lts.viewmodel.ViewModelState;
 import dk.gtz.graphedit.plugins.syntaxes.lts.viewmodel.ViewModelTransition;
+import dk.gtz.graphedit.spi.ISyntaxFactory;
 import dk.gtz.graphedit.tool.IToolbox;
-import dk.gtz.graphedit.view.ISyntaxFactory;
 import dk.gtz.graphedit.view.ModelEditorController;
 import dk.gtz.graphedit.viewmodel.ViewModelEdge;
 import dk.gtz.graphedit.viewmodel.ViewModelVertex;
