+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/petrinet/PNSyntaxFactory.java
@@ -18,9 +18,9 @@ import dk.gtz.graphedit.plugins.syntaxes.petrinet.view.TransitionController;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelArc;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelPlace;
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelTransition;
+import dk.gtz.graphedit.spi.ISyntaxFactory;
 import dk.gtz.graphedit.tool.IToolbox;
 import dk.gtz.graphedit.tool.Toolbox;
-import dk.gtz.graphedit.view.ISyntaxFactory;
 import dk.gtz.graphedit.view.ModelEditorController;
 import dk.gtz.graphedit.viewmodel.ViewModelEdge;
 import dk.gtz.graphedit.viewmodel.ViewModelVertex;
