+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/petrinet/view/ArcController.java
@@ -6,10 +6,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelArc;
+import dk.gtz.graphedit.spi.ISyntaxFactory;
 import dk.gtz.graphedit.tool.ITool;
-import dk.gtz.graphedit.view.EdgeController;
-import dk.gtz.graphedit.view.ISyntaxFactory;
 import dk.gtz.graphedit.util.BindingsUtil;
+import dk.gtz.graphedit.view.EdgeController;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
 import dk.gtz.graphedit.viewmodel.ViewModelProjectResource;
 import javafx.beans.property.DoubleProperty;
