+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/petrinet/view/PlaceController.java
@@ -7,7 +7,7 @@ import org.slf4j.LoggerFactory;
 
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelPlace;
 import dk.gtz.graphedit.tool.ITool;
-import dk.gtz.graphedit.view.ISyntaxFactory;
+import dk.gtz.graphedit.spi.ISyntaxFactory;
 import dk.gtz.graphedit.view.VertexController;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
 import dk.gtz.graphedit.viewmodel.ViewModelGraph;
