+++ std/src/main/java/dk/gtz/graphedit/plugins/syntaxes/petrinet/view/TransitionController.java
@@ -6,8 +6,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import dk.gtz.graphedit.plugins.syntaxes.petrinet.viewmodel.ViewModelTransition;
+import dk.gtz.graphedit.spi.ISyntaxFactory;
 import dk.gtz.graphedit.tool.ITool;
-import dk.gtz.graphedit.view.ISyntaxFactory;
 import dk.gtz.graphedit.view.VertexController;
 import dk.gtz.graphedit.viewmodel.ViewModelEditorSettings;
 import dk.gtz.graphedit.viewmodel.ViewModelGraph;
